Created pasing and returning functions
Functions to pass and return each of the classes by value, reference,
and pointer. Each of the pass functions assigns the parameter to a local
object created on the stack. Each of the return functions returns a
local object created on the stack.

GCC users must compile with the -fpermissive flag, or the code won't
compile (for the above reasons).